working on the client update for now

+++ src/main/java/com/example/smartshop/Repositories/ClientRepository.java
@@ -4,6 +4,8 @@ import com.example.smartshop.Models.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.nio.file.LinkOption;
+
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
 }
